README.md
Looking for an easy way to maximize your crypto earnings in 2025? The XT.com referral code AYMOAQ unlocks $2888 in signup bonuses and up to a 50% discount on trading fees. Whether you're a beginner or a seasoned trader, this is the perfect time to take advantage of XT.com’s new user rewards.